V1

+++ nname/nname/Program.cs
@@ -11,53 +11,64 @@ namespace nname
     {
         static void Main(string[] args)
         {
-            var path = args[0];
+            try
+            {
+                if (args.Length < 1)
+                {
+                    Console.WriteLine("Drag a DLL to this EXE");
+                    return;
+                }
+                var path = args[0];
 
 
-            var pathfix = Path.Combine(Path.GetDirectoryName(path), "Fixed");
-            Directory.CreateDirectory(pathfix);
-            var fileName = Path.GetFileName(path);
-            pathfix = Path.Combine(pathfix, fileName);
-            File.Delete(pathfix);
+                var pathfix = Path.Combine(Path.GetDirectoryName(path), "Fixed");
+                Directory.CreateDirectory(pathfix);
+                var fileName = Path.GetFileName(path);
+                pathfix = Path.Combine(pathfix, fileName);
+                File.Delete(pathfix);
 
-            RType.rnamespace = fileName.Substring(0, fileName.IndexOf('.')) + ".ByteCodes";
-
-
-            ModuleDefMD module = ModuleDefMD.Load(path);
-            Console.WriteLine(module.Assembly);
+                RType.rnamespace = fileName.Substring(0, fileName.IndexOf('.')) + ".ByteCodes";
 
 
-            AssemblyResolver asmResolver = new AssemblyResolver();
-            ModuleContext modCtx = new ModuleContext(asmResolver);
-            asmResolver.DefaultModuleContext = modCtx;
-            asmResolver.EnableTypeDefCache = true;
-
-            module.Context = modCtx;
-            module.Context.AssemblyResolver.AddToCache(module);
+                ModuleDefMD module = ModuleDefMD.Load(path);
+                Console.WriteLine(module.Assembly);
 
 
-            RInstructions rinstructions = new RInstructions(module);
+                AssemblyResolver asmResolver = new AssemblyResolver();
+                ModuleContext modCtx = new ModuleContext(asmResolver);
+                asmResolver.DefaultModuleContext = modCtx;
+                asmResolver.EnableTypeDefCache = true;
 
-            new RField().Apply(module);
-            new RProperty().Apply(module);
-            new RMethod().Apply(module);
+                module.Context = modCtx;
+                module.Context.AssemblyResolver.AddToCache(module);
 
-            new RType().Apply(module);
 
-            rinstructions.Apply();
+                RInstructions rinstructions = new RInstructions(module);
 
-            module.Write(pathfix);
-            module.Dispose();
+                new RField().Apply(module);
+                new RProperty().Apply(module);
+                new RMethod().Apply(module);
 
-            var f = File.OpenRead(pathfix);
-            MemoryStream ms = new MemoryStream();
-            f.CopyTo(ms);
-            f.Dispose();
-            ms.Position = 0;
-            ModuleDefinition cmodule = ModuleDefinition.ReadModule(ms);
-            cmodule.Write(File.Create(pathfix));
-            Helper.WriteLine(path + " -> " + pathfix);
+                new RType().Apply(module);
 
+                rinstructions.Apply();
+
+                module.Write(pathfix);
+                module.Dispose();
+
+                var f = File.OpenRead(pathfix);
+                MemoryStream ms = new MemoryStream();
+                f.CopyTo(ms);
+                f.Dispose();
+                ms.Position = 0;
+                ModuleDefinition cmodule = ModuleDefinition.ReadModule(ms);
+                cmodule.Write(File.Create(pathfix));
+                Helper.WriteLine(path + " -> " + pathfix);
+            }
+            finally
+            {
+                Console.ReadLine();
+            }
         }
 
     }
